+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ## [3.9.1905.0 - unreleased]
 
 ### Added
+- Support to export terms with a specific lcid in TaxonomyExtensions.ExportTermSet
 
 ### Changed